+++ src/biodwh2-core/src/main/java/de/unibi/agbi/biodwh2/core/model/graph/PathMappingDescription.java
@@ -9,7 +9,8 @@ public final class PathMappingDescription {
         INDICATES,
         CONTRAINDICATES,
         INDUCES,
-        INTERACTS
+        INTERACTS,
+        CODES_FOR
     }
 
     private final String type;